% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMDS.R
\name{BMDS}
\alias{BMDS}
\title{Run Bayesian Multidimensional Scaling}
\usage{
BMDS(distances, max_dim, parallel = FALSE, cores)
}
\arguments{
\item{distances}{A (symmetric) matrix of distance values returned by \code{dist}, or a full matrix of values.}

\item{max_dim}{The maximum dimension to be estimated by the algorithm. The BMDS algorithm  will be
run for all dimensions between 1 and \code{max_dim}, and the dimension resulting in the lowest MDSIC value will be chosen as the
final object representation.}

\item{parallel}{Boolean indicating whether or not parallel computing should be utilized when running
multiple MCMC chains. If TRUE, the number of cores must be specified.}

\item{cores}{The number of CPU cores to be used when parallel is set to TRUE.}
}
\value{
A list containing the following items.

\item{X}{A list of matrices, with each element corresponding to a different dimensional representation of the objects.}

\item{sigma_sq}{A list of estimated measurement errors, with each element of the list corresponding to the representation of the objects given in the returned list \code{X}.}

\item{MDSIC}{A vector of Multidimensional Scaling Information Criterion values (MDSIC). Each element of the vector corresponds to the representation of objects given in the returned list \code{X}.}
}
\description{
Represent distances between objects in $p$-dimensional Euclidean space.
}
\examples{
BMDS(cityDIST, max_dim = 6)
}
